MER-301 - added layout for forum (side panels as component) + css fixes
