+++ venus/src/pages/forum/Forum.tsx
@@ -1,32 +1,20 @@
-import { useInfiniteQuery, useQuery } from "@tanstack/react-query";
+import { useInfiniteQuery } from "@tanstack/react-query";
 import Error from "../../components/error/Error.jsx";
-import { fetchPaginatedPosts, fetchCategoriesAndTags } from "../../http/posts";
+import { fetchPaginatedPosts } from "../../http/posts";
 import React, { useEffect, useRef, useState } from "react";
-import AddPostButton from "./components/AddPostButton";
-import ForumSearchBar from "./components/ForumSearchBar";
-import ForumCategoriesTagsPanel from "./categories-and-tags/ForumCategoriesTagsPanel";
-import RightPanel from "./components/RightPanel";
-import ForumFormModal from "./components/ForumFormModal";
-import { useBoolean } from "../../hooks/useBoolean";
-import { useSelector } from "react-redux";
-import { RootState } from "../../redux/store";
-import { notificationAction } from "../../redux/notification";
 import ForumPostPage from "../../model/interface/forum/forumPostPage";
-import useDispatchTyped from "../../hooks/useDispatchTyped";
 import { ForumPostSortOption } from "../../model/enum/forum/forumPostSortOption";
 import ForumPostList from "./components/ForumPostList";
+import ForumLayout from "./components/ForumLayout";
+import SkeletonListedForumPost from "./components/SkeletonListedForumPost";
 
 export default function Forum() {
-    const isLogged = useSelector((state: RootState) => state.account.isLogged);
-    const dispatch = useDispatchTyped();
     const loadMoreRef = useRef<HTMLDivElement>(null);
     const [sortOption, setSortOption] = useState<ForumPostSortOption>({
         name: "Newest",
         sortBy: "PUBLISH_DATE",
         sortDirection: "DESC",
     });
-    const [isModalOpen, setIsModalOpenToTrue, setIsModalOpenToFalse] =
-        useBoolean(false);
 
     const {
         data: postPage,
@@ -47,16 +35,6 @@ export default function Forum() {
         initialPageParam: 0,
     });
 
-    const {
-        data: categoriesAndTags,
-        isLoading: isCatTagsLoading,
-        isError: isCatTagsError,
-        error: catTagsError,
-    } = useQuery({
-        queryKey: ["categoriesAndTags"],
-        queryFn: () => fetchCategoriesAndTags(),
-    });
-
     useEffect(() => {
         const target = loadMoreRef.current;
         if (!target || !hasNextPage) return;
@@ -75,61 +53,37 @@ export default function Forum() {
         };
     }, [hasNextPage, fetchNextPage, sortOption, isFetchingNextPage]);
 
-    const handleAddPostClick = () => {
-        if (isLogged) {
-            setIsModalOpenToTrue();
-        } else {
-            dispatch(
-                notificationAction.addInfo({
-                    message: "Login to create posts.",
-                }),
-            );
-        }
-    };
+    const posts = postPage?.pages.flatMap(
+        (page: ForumPostPage) => page.content ?? [],
+    );
+
+    if (isPostPageLoading) {
+        return (
+            <ForumLayout>
+                <div className="mt-14 min-w-2xl">
+                    {Array.from({ length: 10 }).map((_, i) => (
+                        <SkeletonListedForumPost key={i} />
+                    ))}
+                </div>
+            </ForumLayout>
+        );
+    }
 
     if (isPostPageError) {
         return <Error error={postPageError} />;
     }
 
-    const posts = postPage?.pages.flatMap(
-        (page: ForumPostPage) => page.content ?? [],
-    );
-
     return (
         <>
-            <div className="dark:bg-darkBg dark:text-darkText text-lightText bg-lightBg min-h-screen w-full">
-                <div className="mx-auto flex w-full max-w-6xl flex-row gap-4">
-                    <div className="sticky-forum-panel">
-                        <AddPostButton onClick={handleAddPostClick} />
-
-                        <ForumCategoriesTagsPanel
-                            data={categoriesAndTags}
-                            isLoading={isCatTagsLoading}
-                            isError={isCatTagsError}
-                            error={catTagsError}
-                        />
-                    </div>
-
-                    <ForumPostList
-                        posts={posts}
-                        sortOption={sortOption}
-                        onSortChange={setSortOption}
-                        loadMoreRef={loadMoreRef}
-                        isFetchingNextPage={isFetchingNextPage}
-                    />
-
-                    <div className="sticky-forum-panel">
-                        <ForumSearchBar />
-                        <RightPanel />
-                    </div>
-                    <ForumFormModal
-                        onClose={setIsModalOpenToFalse}
-                        isOpen={isModalOpen}
-                        categories={categoriesAndTags?.categories ?? []}
-                        tags={categoriesAndTags?.tags ?? []}
-                    />
-                </div>
-            </div>
+            <ForumLayout>
+                <ForumPostList
+                    posts={posts}
+                    sortOption={sortOption}
+                    onSortChange={setSortOption}
+                    loadMoreRef={loadMoreRef}
+                    isFetchingNextPage={isFetchingNextPage}
+                />
+            </ForumLayout>
         </>
     );
 }
